# travel-ecosystem-backend/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Trip Planner & Route Optimization API
  description: |
    Complete API specification for the mobile-first trip planning flow:
    - Attraction selection
    - Route optimization (TSP/RAPTOR algorithms)
    - Multimodal transport options (GTFS + Google Directions)
    - PDF itinerary generation
  version: 1.0.0
  contact:
    name: API Support
    email: support@tripplanner.com
servers:
  - url: http://localhost:3001/api/v1
    description: Discovery Engine (Development)
  - url: http://localhost:3007/api
    description: Route Optimizer (Development)
  - url: http://localhost:3008/transport
    description: Transportation Service (Development)
  - url: http://localhost:3009/api
    description: PDF Service (Development)
...
paths:
...
  /optimize-route:
    post:
      tags: [route-optimization]
      summary: Optimize route between attractions
      description: |
        Uses advanced algorithms (TSP, RAPTOR, Genetic, etc.) to find the optimal
        visiting order for selected attractions. Returns order only, not transport details.
      operationId: optimizeRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizeRouteRequest'
      responses:
        '200':
          description: Optimized route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizeRouteResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    OptimizeRouteRequest:
      type: object
      required:
        - places
        - constraints
        - options
      properties:
        userId:
          type: string
          format: uuid
          nullable: true
        places:
          type: array
          minItems: 2
          items:
            type: object
            required: [id, name, lat, lng]
            properties:
              id:
                type: string
              name:
                type: string
              lat:
                type: number
                format: double
              lng:
                type: number
                format: double
              imageUrl:
                type: string
              priority:
                type: integer
              visitDuration:
                type: integer
        constraints:
          type: object
          required: [travelTypes]
          properties:
            startLocation:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
            startTime:
              type: string
              format: date-time
            timeBudgetMinutes:
              type: integer
            travelTypes:
              type: array
              items:
                type: string
                enum: [WALKING, CYCLING, PUBLIC_TRANSPORT, DRIVE, E_SCOOTER]
            budget:
              type: number
        options:
          type: object
          required: [includeRealtimeTransit]
          properties:
            includeRealtimeTransit:
              type: boolean
            algorithm:
              type: string
              enum: [RAPTOR, TSP_2OPT, ACO, GREEDY, GENETIC, CHRISTOFIDES, auto]
              default: auto
    OptimizeRouteResponse:
      type: object
      required:
        - jobId
        - optimizedOrder
        - estimatedDurationMinutes
        - totalDistanceMeters
      properties:
        jobId:
          type: string
          format: uuid
        optimizedOrder:
          type: array
          items:
            type: object
            properties:
              placeId:
                type: string
              seq:
                type: integer
        estimatedDurationMinutes:
          type: integer
        totalDistanceMeters:
          type: integer
        notes:
          type: string
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []